The Great Cleanup
File structure, naming conventions, syntax, and more
